Update cache-store.js
Reset newSize variable after call to cleanCache since cleanCache modifies this.size.  Without this change the stored this.size doesn't accurately represent the cache size.  So when the cache becomes full and the process to clear out old items starts this.size never decrements, resulting in an empty cache and no ability to write new entires.